requirements.txt
aiogram==2.25.1
python-dotenv~=1.0.0
peewee~=3.16.0
requests~=2.31.0
psycopg2-binary==2.9.6
